refactor: domain 내에 car 패키지 추가 생성, 클래스 분리
